+++ src/axolotl/cli/main.py
@@ -2,7 +2,6 @@
 
 # pylint: disable=redefined-outer-name
 
-import os
 import subprocess  # nosec B404
 import tempfile
 from pathlib import Path
@@ -26,6 +25,7 @@ from axolotl.cli.utils import (
     add_options_from_config,
     add_options_from_dataclass,
     build_command,
+    execute_training,
     fetch_from_github,
     filter_none_kwargs,
 )
@@ -117,7 +117,7 @@ def train(
     # Process each configuration
     for cfg_file in config_files:
         try:
-            _execute_training(cfg_file, effective_launcher, cloud, kwargs)
+            execute_training(cfg_file, effective_launcher, cloud, kwargs)
         except subprocess.CalledProcessError as exc:
             LOG.error(f"Failed to train/fine-tune config '{cfg_file}': {exc}")
             if not sweep:
@@ -147,82 +147,6 @@ def _generate_config_files(config: str, sweep: str | None) -> list[str]:
             config_files.append(str(temp_config_path))
 
     return config_files
-
-
-def _execute_training(
-    cfg_file: str, launcher: str | None, cloud: str | None, kwargs: dict
-) -> None:
-    """Execute training with the given configuration."""
-    if cloud:
-        _execute_cloud_training(cloud, cfg_file, launcher, kwargs)
-    elif launcher:
-        _execute_launcher_training(cfg_file, launcher, kwargs)
-
-
-def _execute_cloud_training(
-    cloud: str, cfg_file: str, launcher: str | None, kwargs: dict
-) -> None:
-    """Execute training via cloud launcher."""
-    from axolotl.cli.cloud import do_cli_train
-
-    accelerate = launcher == "accelerate" if launcher else False
-    cwd = os.getcwd() if launcher else None
-
-    do_cli_train(
-        cloud_config=cloud,
-        config=cfg_file,
-        accelerate=accelerate,
-        cwd=cwd,
-        **kwargs,
-    )
-
-
-def _execute_launcher_training(cfg_file: str, launcher: str, kwargs: dict) -> None:
-    """Execute training via specified launcher."""
-    if launcher == "accelerate":
-        _execute_accelerate_training(cfg_file, kwargs)
-    elif launcher == "torchrun":
-        _execute_torchrun_training(cfg_file, kwargs)
-    elif launcher == "python":
-        _execute_python_training(cfg_file, kwargs)
-
-
-def _execute_accelerate_training(cfg_file: str, kwargs: dict) -> None:
-    """Execute training via accelerate launcher."""
-    launcher_args = []
-
-    # Extract launcher-specific arguments
-    if "main_process_port" in kwargs:
-        main_process_port = kwargs.pop("main_process_port")
-        launcher_args.extend(["--main_process_port", str(main_process_port)])
-
-    if "num_processes" in kwargs:
-        num_processes = kwargs.pop("num_processes")
-        launcher_args.extend(["--num_processes", str(num_processes)])
-
-    base_cmd = ["accelerate", "launch"] + launcher_args + ["-m", "axolotl.cli.train"]
-    if cfg_file:
-        base_cmd.append(cfg_file)
-
-    cmd = build_command(base_cmd, kwargs)
-    subprocess.run(cmd, check=True)  # nosec B603
-
-
-def _execute_torchrun_training(cfg_file: str, kwargs: dict) -> None:
-    """Execute training via torchrun launcher."""
-    base_cmd = ["torchrun", "-m", "axolotl.cli.train"]
-    if cfg_file:
-        base_cmd.append(cfg_file)
-
-    cmd = build_command(base_cmd, kwargs)
-    subprocess.run(cmd, check=True)  # nosec B603
-
-
-def _execute_python_training(cfg_file: str, kwargs: dict) -> None:
-    """Execute training via python launcher."""
-    from axolotl.cli.train import do_cli
-
-    do_cli(config=cfg_file, **kwargs)
 
 
 @cli.command()
